+++ src/electron/server.ts
@@ -13,12 +13,7 @@ import {
   getScreenshotImagePathsForDate,
   readFile,
 } from "./files";
-import {
-  buildMinuteSlotsPastWeek,
-  extractKeylogAppSwitchMinuteSet,
-  renderHealthHtml,
-  scanScreenshotSummaryMinuteSet,
-} from "./health";
+import { buildHealthPastWeekHtml } from "./health";
 import { allEndpoints } from "../constants/endpoints";
 
 const transports: { [sessionId: string]: StreamableHTTPServerTransport } = {};
@@ -462,69 +457,13 @@ export function startLocalServer(port = 8765): http.Server {
 
       case "/health": {
         try {
-          const nowTimestampMs = Date.now();
-          const sinceTimestampMs = nowTimestampMs - 7 * 24 * 60 * 60 * 1000;
-          const minuteSlots = buildMinuteSlotsPastWeek(nowTimestampMs);
-
-          const keylogDates: Date[] = [];
-          const today = new Date();
-          for (let i = 0; i < 7; i++) {
-            const date = new Date();
-            date.setDate(today.getDate() - i);
-            keylogDates.push(date);
-          }
-
-          const keylogRawMinutes = new Set<number>();
-          const keylogProcessedMinutes = new Set<number>();
-          for (const keylogDate of keylogDates) {
-            try {
-              const rawPath = getKeyLogFileForDate(keylogDate, "");
-              const rawText = await readFile(rawPath);
-              const rawMinutesForFile = extractKeylogAppSwitchMinuteSet(
-                rawText,
-                sinceTimestampMs,
-                nowTimestampMs,
-              );
-              for (const minuteBucket of rawMinutesForFile) {
-                keylogRawMinutes.add(minuteBucket);
-              }
-            } catch {
-              void 0;
-            }
-            try {
-              const processedPath = getKeyLogFileForDate(
-                keylogDate,
-                "processed.chronological.",
-              );
-              const processedText = await readFile(processedPath);
-              const processedMinutesForFile = extractKeylogAppSwitchMinuteSet(
-                processedText,
-                sinceTimestampMs,
-                nowTimestampMs,
-              );
-              for (const minuteBucket of processedMinutesForFile) {
-                keylogProcessedMinutes.add(minuteBucket);
-              }
-            } catch {
-              void 0;
-            }
-          }
-
-          const screenshotSummaryMinutes = await scanScreenshotSummaryMinuteSet({
+          const html = await buildHealthPastWeekHtml({
             userDataPath,
-            sinceTimestampMs,
-            nowTimestampMs,
+            readFile,
+            getKeyLogFileForDate,
           });
-
           res.writeHead(200, { "Content-Type": "text/html" });
-          res.end(
-            renderHealthHtml({
-              minuteSlots,
-              keylogRawMinutes,
-              keylogProcessedMinutes,
-              screenshotSummaryMinutes,
-            }),
-          );
+          res.end(html);
         } catch {
           res.writeHead(500, { "Content-Type": "text/plain" });
           res.end("Failed to render /health.");
